Move Spotify class out to it's own module
