+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup
 setup(
     name="spotipy-utils",
     version="1.0",
-    packages=["utils", "utils.commands"],
+    packages=["utils", "utils.commands", "utils.internal"],
     include_package_data=True,
     install_requires=["click"],
     entry_points="""
